libc/math: Fix wrong exp*() api usage
> Use correct exp*() api.

+++ lib/libc/math/lib_erf.c
@@ -88,6 +88,6 @@ double erf(double x)
 
 	sign = (x >= 0 ? 1 : -1);
 	t = 1.0 / (1.0 + p * x);
-	return sign * (1.0 - (((((a5 * t + a4) * t) + a3) * t + a2) * t + a1) * t * expf(-x * x));
+	return sign * (1.0 - (((((a5 * t + a4) * t) + a3) * t + a2) * t + a1) * t * exp(-x * x));
 }
 #endif
