+++ lib/libc/math/lib_logf.c
@@ -66,7 +66,7 @@ float logf(float x)
 
 	while (y > y_old + epsilon || y < y_old - epsilon) {
 		y_old = y;
-		ey = exp(y);
+		ey = expf(y);
 		y -= (ey - x) / ey;
 
 		if (y > 700.0) {
